test s3 image uploads

+++ src/pages/api/posts/delete.ts
@@ -3,6 +3,14 @@ import { prisma } from '@backend/index';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
     try {
+        if (req.method !== 'GET') {
+            res.setHeader('Allow', 'GET');
+            res.status(405).json({
+                data: null,
+                error: 'Method Not Allowed',
+            });
+            return;
+        }
         const result = await prisma.post.delete({
             where: {
                 id: Number(req?.body?.id),
